[feat] launch_check_model done for titanic
