+++ bin/launch_training.py
@@ -1,14 +1,12 @@
 #!/usr/bin/env python3
 
 import argparse
-from src.learner.raytune_learner import TuneTrainWrapper as StimulusTuneTrainWrapper
-import src.data.experiments as exp
-import ray.tune as tune
-from ray import train, tune
-import ray.tune.schedulers as schedulers
 import json
 import os
 import importlib.util
+
+from src.learner.raytune_learner import TuneTrainWrapper as StimulusTuneTrainWrapper
+from launch_utils import import_class_from_file, get_experiment
 
 def get_args():
 
@@ -24,8 +22,6 @@ def get_args():
 
     args = parser.parse_args()
     return args
-
-
 
 def import_class_from_file(file_path: str) -> type:
 
@@ -47,12 +43,6 @@ def import_class_from_file(file_path: str) -> type:
     
     # Class not found
     raise ImportError("No class starting with 'Model' found in the file.")
-
-
-def get_experiment(experiment_name: str) -> object:
-    experiment_object = getattr(exp, experiment_name)()
-    return experiment_object
-
 
 def main(config_path: str, model_path: str, data_path: str, json_experiment: str, output: str, best_config_path: str) -> None:
     """
@@ -83,8 +73,7 @@ def main(config_path: str, model_path: str, data_path: str, json_experiment: str
     # Train the model with the best config and best model, aka fine-tuning
     #learner.train()
     # Save the model fine-tuned model
-    #learner.trainer.export_model(output)
-    
+    #learner.trainer.export_model(output)    
 
 if __name__ == "__main__":
     args = get_args()
